September-october commit
- manual parser
- stl refinements
- all language constructs are now Send

+++ src/rule/vec_attrs.rs
@@ -5,7 +5,7 @@ use crate::interner::Tok;
 /// Returns the name, priority and nonzero of the expression if it is
 /// a vectorial placeholder
 #[must_use]
-pub fn vec_attrs(expr: &RuleExpr) -> Option<(Tok<String>, u64, bool)> {
+pub fn vec_attrs(expr: &RuleExpr) -> Option<(Tok<String>, usize, bool)> {
   match expr.value.clone() {
     Clause::Placeh(Placeholder {
       class: PHClass::Vec { prio, nonzero },
